refactor(core): don't worry about return values for informational events
